Override `translate_exception_class` instead of `raw_execute` and `execute`
Starting with Rails 7.1, `raw_execute` may not be called when executing a query. This pull request introduces `with_raw_connection`, and this method is now called when executing a query.

https://github.com/rails/rails/pull/44576/files#diff-460f4e7973c5dd945c51d24df5b0173961190d3645f4e2585fd3003fa1fc0ff7R865

Overriding this method and calling super will result in a LocalJumpError, which is not safe.

In this commit, I tried to call the log by hooking into `translate_exception_class`. This method has been around since 4.2, and there should be no changes to the interface since 6.0.

https://github.com/rails/rails/commit/5e5118aa8ba821af6e615cfd2903b848ff8a9177#diff-460f4e7973c5dd945c51d24df5b0173961190d3645f4e2585fd3003fa1fc0ff7R356